Migrate to Vitest, refactor cache-handler API names
