+++ internal/typescript-config/test.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "display": "Test",
+  "compilerOptions": {
+    "target": "es2022",
+    "lib": ["es2023"],
+
+    "module": "NodeNext",
+    "moduleDetection": "force",
+    "esModuleInterop": true,
+
+    "strict": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    "skipLibCheck": true,
+
+    "noEmit": true
+  }
+}
