--- packages/cache-handler/src/helpers/get-timeout-redis-command-options.ts
@@ -1,13 +1,0 @@
-import { commandOptions } from 'redis';
-
-type CommandOptions = ReturnType<typeof commandOptions>;
-
-export function getTimeoutRedisCommandOptions(
-  timeoutMs: number,
-): CommandOptions {
-  if (timeoutMs === 0) {
-    return commandOptions({});
-  }
-
-  return commandOptions({ signal: AbortSignal.timeout(timeoutMs) });
-}
