+++ packages/cache-handler/src/helpers/is-tag-implicit.ts
@@ -1,0 +1,29 @@
+import { NEXT_CACHE_IMPLICIT_TAG_ID } from '../next-common-types.js';
+
+/**
+ * Determines if a tag is an implicit tag generated by Next.js for internal cache management.
+ *
+ * Implicit tags are automatically generated by Next.js based on the routing hierarchy
+ * to manage cache invalidation for pages and layouts in the App Router. They are prefixed
+ * with `NEXT_CACHE_IMPLICIT_TAG_ID` ('_N_T_') to distinguish them from user-defined tags.
+ *
+ * Implicit tags require special handling during revalidation because they are part of
+ * Next.js's internal caching mechanism. When an implicit tag is revalidated, the actual
+ * cache invalidation is deferred until the next 'get' operation that accesses the cache.
+ *
+ * @param tag - The cache tag string to check
+ *
+ * @returns `true` if the tag is an implicit Next.js tag (starts with '_N_T_'), `false` otherwise
+ *
+ * @example
+ * // Returns true for implicit tags
+ * isTagImplicit('_N_T_/layout')          // true
+ * isTagImplicit('_N_T_/user/[id]/page')  // true
+ *
+ * // Returns false for user-defined tags
+ * isTagImplicit('user-123')              // false
+ * isTagImplicit('product-update')        // false
+ */
+export function isTagImplicit(tag: string): boolean {
+  return tag.startsWith(NEXT_CACHE_IMPLICIT_TAG_ID);
+}
