+++ packages/cache-handler/test/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "@repo/typescript-config/test.json",
+  "include": ["**/*.test.ts", "**/*.spec.ts"]
+}
