+++ packages/cache-handler/tsconfig.json
@@ -6,5 +6,5 @@
     "rootDir": "src"
   },
   "include": ["**/*.ts"],
-  "exclude": ["**/*.test.ts", "node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "test", "vitest.config.ts"]
 }
